+++ src/main/java/baseball/GameDecision.java
@@ -1,8 +1,5 @@
 package baseball;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public enum GameDecision {
     RESTART("1"), // 게임 재시작
     EXIT("2");     // 게임 완전히 종료
